Meetmanager/requirements.txt
gradio
google-api-python-client
google-auth
google-auth-oauthlib
pandas
openpyxl
